esc before save
I was wasting time trying to ^S on a file and nvim actually typing it
out. So I try to ^S, then have to delete ^S, esc, and save. So now <Esc>
is auto done so I can save even in insert mode.
Also removed redundant : with <cmd>, which was doing <cmd>:w instead of
just <cmd>w